third_party/blink/web_tests/http/tests/xmlhttprequest/set-dangerous-headers-expected.txt

CONSOLE ERROR: Refused to set unsafe header "ACCEPT-CHARSET"
CONSOLE ERROR: Refused to set unsafe header "ACCEPT-ENCODING"
CONSOLE ERROR: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-HEADERS"
CONSOLE ERROR: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-METHOD"
CONSOLE ERROR: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-PRIVATE-NETWORK"
CONSOLE ERROR: Refused to set unsafe header "CONNECTION"
CONSOLE ERROR: Refused to set unsafe header "CONTENT-LENGTH"
CONSOLE ERROR: Refused to set unsafe header "COOKIE"
CONSOLE ERROR: Refused to set unsafe header "COOKIE2"
CONSOLE ERROR: Refused to set unsafe header "DATE"
CONSOLE ERROR: Refused to set unsafe header "DNT"
CONSOLE ERROR: Refused to set unsafe header "EXPECT"
CONSOLE ERROR: Refused to set unsafe header "HOST"
CONSOLE ERROR: Refused to set unsafe header "KEEP-ALIVE"
CONSOLE ERROR: Refused to set unsafe header "ORIGIN"
CONSOLE ERROR: Refused to set unsafe header "REFERER"
CONSOLE ERROR: Refused to set unsafe header "TE"
CONSOLE ERROR: Refused to set unsafe header "TRAILER"
CONSOLE ERROR: Refused to set unsafe header "TRANSFER-ENCODING"
CONSOLE ERROR: Refused to set unsafe header "UPGRADE"
CONSOLE ERROR: Refused to set unsafe header "USER-AGENT"
CONSOLE ERROR: Refused to set unsafe header "VIA"
CONSOLE ERROR: Refused to set unsafe header "Proxy-"
CONSOLE ERROR: Refused to set unsafe header "Proxy-test"
CONSOLE ERROR: Refused to set unsafe header "PROXY-FOO"
CONSOLE ERROR: Refused to set unsafe header "Sec-"
CONSOLE ERROR: Refused to set unsafe header "Sec-test"
CONSOLE ERROR: Refused to set unsafe header "SEC-FOO"
Test that setRequestHeader cannot be used to alter security-sensitive headers.

SUCCESS
